refact: move `filter_versions` out of `upgrade_python_package`
It's a function that's used in multiple modules. Move it outside of the
main entrypoint script, so we avoid potential circular imports.
